src/authentication/serializers.py
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer


class UserTokenObtainPairSerializer(TokenObtainPairSerializer):
    @classmethod
    def get_token(cls, user):
        token = super().get_token(user)

        token["slug"] = user.slug
        token["last_task_id"] = user.task_id

        return token
