pyproject.toml
[tool.poetry]
name = "geoapi"
version = "0.1.0"
description = ""
authors = ["Kornel Koszela <kornel.koszela@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
Django = "^3.2"
psycopg2-binary = "^2.9.3"
redis = "^4.2.1"
uritemplate = "^4.1.1"
django-filter = "^21.1"
djangorestframework = "^3.13.1"
channels = "^3.0.4"
channels-redis = "^3.4.0"
celery = {extras = ["redis"], version = "^5.2.3"}

djangorestframework-simplejwt = "^5.1.0"
types-requests = "^2.27.16"
django-extensions = "^3.1.5"
whitenoise = "^6.0.0"

drf-spectacular = {extras = ["sidecar"], version = "^0.22.0"}
uvicorn = "^0.17.6"
requests = "^2.27.1"
coverage = "^6.3.2"
dj-database-url = "^0.5.0"
validators = "^0.18.2"


[tool.poetry.dev-dependencies]
bandit = "^1.7.4"
flake8 = "^4.0.1"
mypy = "^0.942"
pytest = "^7.1.1"
pytest-django = "^4.5.2"
pytest-mock = "^3.7.0"
black = "^22.3.0"
pytest-asyncio = "^0.18.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.bandit]
exclude = "test_*.py"

[tool.black]
line-length = 79
